% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RcmdrPlugin.cpd}
\alias{RcmdrPlugin.cpd}
\alias{RcmdrPlugin.cpd-package}
\title{cpd Plug-In for R Commander}
\description{
This package is an R Commander "plug-in" based on the
\code{cpd} package (see \code{\link{cpd}}). It provides a GUI for fitting a generalized Waring regression model to a
dataset, together with other common regression functions such as Akaike and
Bayesian information criteria, stepwise model selection or add observation
statistics to data. Moreover, it allows for computing the partition of the
variance for the selected \code{GW} model and carrying out a graphical diagnosis
by a simulated envelope of the residuals.
}
\details{
Package: RcmdrPlugin.cpd

Type: Package

Title: R Commander Plug-In for Complex Pearson distributions

Version: 0.0.1

Date: 2019-09-01

Depends: R (>= 3.0.0), cpd (>= 0.1.0), RcmdrMisc (>= 1.0-2)

License: GPL (>= 2)
}
\note{
Some functions cannot run adequately if there are missing data.
}
\seealso{
\code{\link{Rcmdr}}
}
\author{
Silverio Vilchez-Lopez and Maria Jose
Olmo-Jimenez.
Maintainer: Silverio Vilchez-Lopez
}
