% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RcmdrPlugin.cpd}
\alias{RcmdrPlugin.cpd}
\title{cpd Plug-In for R Commander}
\description{
This package is an R Commander "plug-in" based on the
\code{cpd} package (see \code{\link{cpd}}). It provides a GUI for ...
}
\details{
Package: RcmdrPlugin.cpd

Type: Package

Title: R Commander Plug-In for Complex Pearson distributions

Version: 0.0.1

Date: 2022-08-16

Depends: R (>= 4.2.0), cpd (>= 0.3.0), RcmdrMisc (>= 2.7-2)

License: GPL (>= 2)
}
\note{
Some functions cannot run adequately if there are missing data.
}
\seealso{
\code{\link{Rcmdr}}
}
\author{
Silverio Vilchez-Lopez and Maria Jose
Olmo-Jimenez.
Maintainer: Silverio Vilchez-Lopez
}
