% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.cpd.r
\name{RcmdrPlugin.Utility}
\alias{RcmdrPlugin.Utility}
\alias{cbpSamples}
\alias{ctpSamples}
\alias{ebwSamples}
\alias{cbpPlot}
\alias{ctpPlot}
\alias{ebwPlot}
\alias{cbpMass}
\alias{ctpMass}
\alias{ebwMass}
\alias{cbpProbabilities}
\alias{ctpProbabilities}
\alias{ebwProbabilities}
\alias{cbpQuantiles}
\alias{ctpQuantiles}
\alias{ebwQuantiles}
\title{'cpd' Plug-In Utility Functions}
\usage{
cbpSamples()

ctpSamples()

ebwSamples()

cbpPlot()

ctpPlot()

ebwPlot()

cbpMass()

ctpMass()

ebwMass()

cbpProbabilities()

ctpProbabilities()

ebwProbabilities()

cbpQuantiles()

ctpQuantiles()

ebwQuantiles()
}
\value{
Not return value, the effects are shown in 'Rcmdr'  GUI
}
\description{
'cpd' Plug-In Utility Functions
}
