% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcmdrPlugin.cpd.r
\name{RcmdrPlugin.Utility}
\alias{RcmdrPlugin.Utility}
\alias{cbpProbabilities}
\alias{ctpProbabilities}
\alias{cbpQuantiles}
\alias{ctpQuantiles}
\title{cpd Plug-In Utility Functions}
\usage{
cbpProbabilities()

ctpProbabilities()

cbpQuantiles()

ctpQuantiles()
}
\description{
cpd Plug-In Utility Functions
}
